GH-569 Check when to delete old cache data
Only delete the cache for the previous round if at least one question was already played.

+++ classes/teststrategy/progress.php
@@ -837,12 +837,14 @@ class progress implements JsonSerializable {
             $response = catquiz::get_last_response_for_attempt($this->get_usage_id());
             $cache->set($cachekey, $response);
             // Delete the cache entry for the previous number of questions answered.
-            $previouskey = sprintf(
-                'lastresponse_%d_%d',
-                $this->get_usage_id(),
-                $this->get_num_playedquestions() - 1
-            );
-            $cache->delete($previouskey);
+            if ($this->get_num_playedquestions() >= 1) {
+                $previouskey = sprintf(
+                    'lastresponse_%d_%d',
+                    $this->get_usage_id(),
+                    $this->get_num_playedquestions() - 1
+                );
+                $cache->delete($previouskey);
+            }
         }
         if ($response && $response->state === 'gaveup') {
             $response->fraction = 0.0;
